requirements.txt
Flask==3.0.3
Pillow==9.4.0
Flask-Limiter==3.8.0

